requirements.txt
requests
configparser
paho-mqtt